* Applying selected policy to all actions feature was added * Applying selected policy to actions only in a specific category feature was added * Reseting changes made feature was added * Finding the correct grant/revoke status after shuffling choices was fixed * Updating user's main group was fixed * Setting icons of actions correctly after expanding the listview was fixed * The exception which occurs when a user clicks the empty area in user list was fixed * Enabled status of Ask Password check box after clicking a blocked action was fixed * Deleting a user after cancelling the authorization dialog was fixed * Adding a user whose custom user id smaller than 1000 was fixed * Informing the user if a user id is already in use before add operation is triggered feature was added
I think this commit solves below bugs:
BUG:COMMENT:5854
BUG:COMMENT:7851
BUG:COMMENT:16874
BUG:COMMENT:16873

+++ branches/kde/user-manager/corporate2/browser.py
@@ -148,7 +148,10 @@ class BrowseStack(QVBox):
                     if exception:
                         return
                     self.userModified(item.uid)
-                self.mainwidget.link.User.Manager["baselayout"].deleteUser(item.uid, ret == KMessageBox.Yes, async=handler)
+                if ret == KMessageBox.Yes or ret == KMessageBox.No:
+                    self.mainwidget.link.User.Manager["baselayout"].deleteUser(item.uid, ret == KMessageBox.Yes, async=handler)
+                elif ret == KMessageBox.Cancel:
+                    pass
         else:
             item = self.groups.selectedItem()
             if item:
